+++ chevah/compat/tests/normal/test_filesystem.py
@@ -969,13 +969,13 @@ class TestLocalFilesystem(CompatTestCase, FilesystemTestMixin):
         if self.os_name == 'aix':
             count = 3000
             base_timeout = 0.02
-        elif self.os_name == 'hpux':
-            # HP-UX does not allow more than 32765 members in a folder
+        elif self.os_name in ['hpux', 'freebsd', 'openbsd']:
+            # Some OS/FS does not allow more than 32765 members in a folder
             # and the slave is generally slow.
             count = 32000
             base_timeout = 0.1
         else:
-            count = 35000
+            count = 45000
             base_timeout = 0.1
 
         base_segments = self.folderInTemp()
@@ -995,6 +995,7 @@ class TestLocalFilesystem(CompatTestCase, FilesystemTestMixin):
 
         # Iterating at any step will not take long.
         result = []
+        result.append(next(iterator))
         try:
             while True:
                 with self.assertExecutionTime(base_timeout):