+++ chevah/compat/tests/normal/testing/test_testcase.py
@@ -318,7 +318,6 @@ class TestTwistedTestCase(ChevahTestCase):
         def last_call():
             time.sleep(0.2)
             self.called = True
-            time.sleep(0.01)
 
         deferred = threads.deferToThread(last_call)
 
@@ -561,7 +560,14 @@ class TestChevahTestCase(ChevahTestCase):
 
         This is a system test, but socket operations are light.
         """
-        if self.os_name in ['aix', 'solaris', 'osx', 'hpux']:
+        if self.os_name in [
+            'aix',
+            'freebsd',
+            'hpux',
+            'openbsd',
+            'osx',
+            'solaris',
+                ]:
             # On AIX and probably on other Unixes we can only bind on
             # existing fixed IP addressed like 127.0.0.1.
             raise self.skipTest()